Step 2 Good Eno'
Cleared up a bunch of nonsense and got both the accordion and the outline
into KO+template form.

Might add a bit of jQuery to hide the entire "kind of treatment" DIV if
none of that kind's treatments are supported. (jQuery because it's a
pure-view issue and because it's easier to bind to the <change> event
and dig "up" the DOM to the enclosing DIV with jq than it is to build a
KO dependent observable structure to "watch" the kind's childs' supported
attributes.)