Time bands on the calendar and fixed a bug!
Added the ability to display colored time bands on the calendar display
and they move with the scale as you slide the slider. Neat.

Also fixed the nasty NaN bug that caused annoying flicker sometimes while
scrolling. Apparently setting duration (of the redraw-on-slide
transitions) was causing occasional divide-by-zero errors somewhere. It's
all fixed now and scrolling is lovely and smooth. Yay!