New build system, tests, juvume project.
Build System
============

This commit sees the introdcution of an actual build system for our site.  It
is only just compiling the JS to more compact and efficient JS and minifying
the CSS, but it also runs both Google's JS linter and JSLint over the source to
check for errors and style violations.

It's slightly overkill for us at this point but valuabe to get right right from
the [re-]start.

Redo
----

Instead of 'make' or some other build software we are using the near-magical
'Redo' system conceived by Daniel J. Bernstein and implemented by Apenwarr
(whom I worked for at NITI. Brilliant guy.)

This is an extremely simple and powerful suite of tools that leverage the power
of the shell to create sophisticated and flexible build scripts that are easy
to develop and maintain.

https://github.com/apenwarr/redo

"It's like butter."

Build Directory
---------------

All of the build goodness resides and takes place in the new 'build' directory.

A typical build collects and compiles 'site.js' and 'site.css' files in the
'build/out' directory and then copies them to the Django project's
'juvume/static/js' and 'juvume/static/css' directories respectively.

The source for the CSS files are the *.gss files in the 'build/src-static/css'
directory, and the JS are in the 'build/src-static/js' directory.

Closure and JSLint
------------------

We're using Google's Closure suite:

* The CSS compiler transforms and minimizes "GSS" files to CSS, and has
provisions for minifying CSS class names that we can use later.

* The gjslint linter checks a number of things including proper documentation.

* The JS compiler does extensive checking and minification as well as dependency
analysis and "dead code" removal.

* We also use Crockford's JSLint to be extra frosty with our JS goodness. (We run
it using Rhino JS engine.)

Tests
=====

We also now have exactly one test that verifies that you can use the splash
page email form to get an email into our database.

It's written using Selenium and the test case has been exported both as HTML
and Python so you can run it in the browser IDE or from the command line.

New Revision of the Juvume Site
===============================

In order to support a quick revision of the site to accomodate the new build
system and a more systematic development process I've created a new Django
project in the 'juvume' directory.

It currently contains only the 'splash' app but I'll be bringing over the
'look' module starting this evening and we should be serving abd booking
appointments by Thursday.